+++ src/colorant.rs
@@ -52,6 +52,13 @@ mod munsell_matt;
 #[cfg(feature = "munsell")]
 pub use munsell_matt::*;
 
+#[cfg(feature = "cri")]
+mod tcs;
+
+#[cfg(feature = "cri")]
+pub use tcs::{TCS, N_TCS};
+
+
 use std::{
     borrow::Cow,
     ops::{Add, AddAssign, Mul},
